backend/pyproject.toml
[project]
name = "eks-operations-assistant-api"
version = "0.1.0"
description = "REST API for EKS diagnostics and assistant workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.27.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "aiosqlite>=0.20.0",
  "kubernetes>=31.0.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
